mapper: move things around a bit
